Fixed arrow keys removed duplication of code for elden ring macro
