+++ src/main/java/local/macroj/data/macro/eldenring/SelectConsumableSlot.java
@@ -1,58 +1,25 @@
 package local.macroj.data.macro.eldenring;
 
-import com.google.gson.reflect.TypeToken;
-import local.macroj.data.MacroKey;
-import local.macroj.data.macro.Macro;
-import local.macroj.data.macro.eldenring.data.SelectSlot;
 import local.macroj.sender.Sender;
-import lombok.RequiredArgsConstructor;
-import lombok.SneakyThrows;
+import lombok.experimental.SuperBuilder;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
-import java.lang.reflect.Type;
 
 @Component
-@RequiredArgsConstructor
 @Slf4j
 @Scope("prototype")
-public class SelectConsumableSlot implements Macro {
+public class SelectConsumableSlot extends SelectGenericSlot {
 
-    private final InMemoryHandler data;
-    private final Sender sender;
-    private SelectSlot slotInfo;
-
-    @Override
-    public Type getParamsType() {
-        return TypeToken.get(SelectSlot.class).getType();
+    @Autowired
+    public SelectConsumableSlot(InMemoryHandler data, Sender sender) {
+        super(data, sender);
     }
 
     @Override
-    public void setParams(Object param, MacroKey self) {
-        slotInfo = (SelectSlot)param;
-    }
-
-    @SneakyThrows
-    @Override
-    public void run() {
-        int amount = data.selectConsumable(slotInfo.getPosition());
-        for (int i = 0; i < amount; i++) {
-            sender.sendKey(slotInfo.getChangeKey(), 32);
-            Thread.sleep(32);
-        }
-
-        if (slotInfo.getUseKey() != null) {
-            slotInfo.getUseKey().stream().forEach(
-                    use -> {
-                        sender.sendKey(use.getKey(), 64);
-                        try {
-                            Thread.sleep(use.getDelay());
-                        } catch (InterruptedException e) {
-                            e.printStackTrace();
-                        }
-                    }
-            );
-        }
+    protected int selectShift(int position) {
+        return data.selectConsumable(position);
     }
 }
