+++ src/main/java/local/macroj/jna/Keyboard.java
@@ -6,6 +6,7 @@ import com.sun.jna.platform.win32.BaseTSD.ULONG_PTR;
 import com.sun.jna.platform.win32.WinDef.DWORD;
 import com.sun.jna.platform.win32.WinDef.WORD;
 import com.sun.jna.platform.win32.WinUser.INPUT;
+import local.macroj.data.Key;
 
 import static local.macroj.jna.Windows.IS_MACRO;
 
@@ -31,59 +32,38 @@ public class Keyboard {
 	}
 
 	/**
-	 * Sends a key-down input followed by a key-up input for the given character
-	 * value vkCode.
-	 * 
-	 * @param vkCode
-	 */
-	public static void pressKey(int vkCode, int scanCode, boolean allowRecursive) {
-		INPUT input = new INPUT();
-		input.type = new DWORD(INPUT.INPUT_KEYBOARD);
-		input.input.setType("ki");
-		input.input.ki.wScan = new WORD(scanCode);
-		input.input.ki.time = new DWORD(0);
-		input.input.ki.dwExtraInfo = new ULONG_PTR(allowRecursive ? 0 : IS_MACRO);
-		input.input.ki.wVk = new WORD(vkCode);
-		input.input.ki.dwFlags = new DWORD(KEYEVENTF_KEYDOWN);
-		User32.INSTANCE.SendInput(new DWORD(1), (INPUT[]) input.toArray(1), input.size());
-		input.input.ki.wVk = new WORD(vkCode);
-		input.input.ki.dwFlags = new DWORD(KEYEVENTF_KEYUP);
-		User32.INSTANCE.SendInput(new DWORD(1), (INPUT[]) input.toArray(1), input.size());
-	}
-
-	/**
 	 * Sends a key-down input for the given character value vkCode.
 	 *
-     * @param vkCode
+     * @param key
      * @param allowRecursive
      */
-	public static void sendKeyDown(int vkCode, int scanCode, boolean allowRecursive) {
+	public static void sendKeyDown(Key key, boolean allowRecursive) {
 		INPUT input = new INPUT();
 		input.type = new DWORD(INPUT.INPUT_KEYBOARD);
 		input.input.setType("ki");
-		input.input.ki.wScan = new WORD(scanCode);
+		input.input.ki.wScan = new WORD(key.getScanCode());
 		input.input.ki.time = new DWORD(0);
 		input.input.ki.dwExtraInfo = new ULONG_PTR(allowRecursive ? 0 : IS_MACRO);
-		input.input.ki.wVk = new WORD(vkCode);
-		input.input.ki.dwFlags = new DWORD(KEYEVENTF_KEYDOWN);
+		input.input.ki.wVk = new WORD(key.getKeyCode());
+		input.input.ki.dwFlags = new DWORD(key.isExtended() ? KEYEVENTF_KEYDOWN + 1 : KEYEVENTF_KEYDOWN);
 		User32.INSTANCE.SendInput(new DWORD(1), (INPUT[]) input.toArray(1), input.size());
 	}
 
 	/**
 	 * Sends a key-up input for the given character value vkCode.
 	 *
-	 * @param vkCode
+	 * @param key
 	 * @param allowRecursive
 	 */
-	public static void sendKeyUp(int vkCode, int scanCode, boolean allowRecursive) {
+	public static void sendKeyUp(Key key, boolean allowRecursive) {
 		INPUT input = new INPUT();
 		input.type = new DWORD(INPUT.INPUT_KEYBOARD);
 		input.input.setType("ki");
-		input.input.ki.wScan = new WORD(scanCode);
+		input.input.ki.wScan = new WORD(key.getScanCode());
 		input.input.ki.time = new DWORD(0);
 		input.input.ki.dwExtraInfo = new ULONG_PTR(allowRecursive ? 0 : IS_MACRO);
-		input.input.ki.wVk = new WORD(vkCode);
-		input.input.ki.dwFlags = new DWORD(KEYEVENTF_KEYUP);
+		input.input.ki.wVk = new WORD(key.getKeyCode());
+		input.input.ki.dwFlags = new DWORD(key.isExtended() ? KEYEVENTF_KEYUP + 1 : KEYEVENTF_KEYUP);
 		User32.INSTANCE.SendInput(new DWORD(1), (INPUT[]) input.toArray(1), input.size());
 	}
 }
